api/base_requirements.txt
# fmt: off
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
bcrypt==4.2.0
pydantic==2.9.2
pydantic[email]==2.9.2
sqlalchemy==2.0.35
python-jose==3.3.0
passlib==1.7.4
langgraph==0.2.53
langchain-community==0.3.7
langchain-openai==0.2.9
langchain-qdrant==0.2.0
langchain-postgres==0.0.12
psycopg==3.2.3
psycopg-pool==3.2.4
psycopg2-binary==2.9.10 